requirements.txt
streamlit==0.71.0
typer==0.3.2
Pillow==8.0.1
pandas==1.1.4
joblib==0.17.0